feat(canvas): enhance DrawingCanvas with reload functionality and add GadgetPropertiesPanel for editing properties
